Fix compiler warnings (only in HEAD), same style of silence as in 5.3
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@280602 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/iconv/iconv.c
@@ -753,15 +753,15 @@ static php_iconv_err_t _php_iconv_substr(smart_str *pretval,
 		}
 	}
 
-	if(len > total_len) {
+	if((unsigned int) len > total_len) {
 		len = total_len;
 	}
 
-	if (offset >= total_len) {
+	if ((unsigned int) offset >= total_len) {
 		return PHP_ICONV_ERR_SUCCESS;
 	}
 
-	if ((offset + len) > total_len) {
+	if ((unsigned int) (offset + len) > total_len) {
 		/* trying to compute the length */
 		len = total_len - offset;
 	}
